+++ src/guild_log_analysis/analysis/bosses/__init__.py
@@ -1,7 +1,9 @@
 """Boss analysis implementations package."""
 
 from .one_armed_bandit import OneArmedBanditAnalysis
+from .sprocketmonger_lockenstock import SprocketmongerLockenstockAnalysis
 
 __all__ = [
     "OneArmedBanditAnalysis",
+    "SprocketmongerLockenstockAnalysis",
 ]